Modified MakeMsi.bat so it takes the license from ext/subversion/subversion/LICENSE, applies the Text2RTF tool to it, and generates an RTF version of that, puts it in src/TortoiseSVNSetup/include for the setup package to scoop up.
Modified StructureFragment.wxs so it takes the license from the external, not a copy.

These two changes mean any changes to Subversion's license will be pulled into our installer as part of the automatic build process.  They also mean that I can delete the copy and the rtf version of the copy from the setup include directory.

git-svn-id: https://svn.code.sf.net/p/tortoisesvn/code/trunk@4113 154be670-7969-47af-9ab0-c6fb378d2295