Don't set the process priority of the cache! Instead, set the priority of the different threads. Especially on Vista, the priority PROCESS_MODE_BACKGROUND_BEGIN renders the cache completely useless: with that priority, the process can take a *long* time until it gets its share of processor time - even on systems that don't have a high load, it can take 10 minutes or more where the process doesn't get a single CPU cycle! Even exiting the cache process from its tray icon context menu took more than 30 seconds!!
Instead, leave the process priority normal and reduce the thread priority for the folder crawler and the shell notifyer. On Vista, the folder crawler thread will enter  THREAD_MODE_BACKGROUND after being woken up until the crawling is over. But while it is in sleep/wait state for a wakeup notification, the thread priority is restored to IDLE to make sure it gets woken up.

Other threads like the directory watcher keep their priority.

git-svn-id: https://svn.code.sf.net/p/tortoisesvn/code/trunk@11421 154be670-7969-47af-9ab0-c6fb378d2295