Patch from Dmitry: This behavior is only reproduced for deleting or inserting lines, not for changing M-to-M. Just delete or insert one line and scroll down so that the line in question is obscured by the view top. If a line was inserted the view scrolls up on each reload. If a line was deleted the view scrolls down.
Here's why:

When a line is inserted CMainFrame::LoadViews() invokes

    nOldLineNumber = m_pwndLeftView->m_pViewData->GetLineNumber(m_pwndLeftView->m_nTopLine);

the current top line M corresponds to line M-1 in the original file, this M-1 value is later used as the scroll target.

The problem occurs for a similar reason when a line is deleted. The assignment in the biginng maps the top line M to line M in the original file, then this value is passed into

    n = m_pwndRightView->m_pViewData->FindLineNumber(n);

and is mapped onto line M+1 in the resulting file.


git-svn-id: https://svn.code.sf.net/p/tortoisesvn/code/trunk@18738 154be670-7969-47af-9ab0-c6fb378d2295