Make the normal cleanup command only run the cleanup. To refresh the shell overlays, shift-right-click to get the new "cleanup and refresh" menu which does what the cleanup command did before.
Reason for splitting this up:
With WC-NG and the single-db design, we don't have a simple check anymore whether a directory is versioned (i.e., has an .svn dir). That means that we always recurse into all folders, even unversioned ones. And that can take way too much time for big working copies with even bigger unversioned folders in it.

git-svn-id: https://svn.code.sf.net/p/tortoisesvn/code/trunk@19990 154be670-7969-47af-9ab0-c6fb378d2295