patch from ps@pks.im:
When the registry key "HKCU\Software\TortoiseSVN\CommitReopen" is set to "1", then the Copy dialog will reopen in case the commit failed. If this happens and the repository contains externals, then the list of externals will display all externals multiple times, as the externals vector is not getting cleared before it's being repopulated.

This change fixes that by first clearing the externals vector. 

git-svn-id: http://svn.osdn.net/svnroot/tortoisesvn/trunk@28219 154be670-7969-47af-9ab0-c6fb378d2295