patch from Dave Lawrence (slightly modified by me): https://groups.google.com/g/tortoisesvn/c/gYGra7Bw_WU/m/BMXL6DwfBAAJ
add a “Show log of repo at revision…” command to the Explorer extended context menu. Selecting it on any SVN-controlled file or folder will show a dialog asking for a revision number, then show the log starting at that revision. It always shows the log of the root folder, ignoring the specific file(s) or folder(s) that were selected in Explorer.

git-svn-id: http://svn.osdn.net/svnroot/tortoisesvn/trunk@29527 154be670-7969-47af-9ab0-c6fb378d2295